Organize setting tabs seperate folder
